refactor: Extract add dataset
